LayoutTests/platform/gtk/fast/css/getComputedStyle/marginComputedStyle-expected.txt

Test
width: 392px
height: 25px
margin-left: 186px
margin-right: 186px
margin-top: 0px
margin-bottom: 0px
padding-left: 0px
left: auto
